// src/modules/employee-document/dto/enums/employee-document.status.enum.ts
export enum EMPLOYEE_DOCUMENT_STATUS {
  UPLOADED_FOR_EMPLOYEE = 'STATUS_UPLOADED_FOR_EMPLOYEE',
  WAITING_FILE = 'STATUS_WAITING_FILE',
  WAITING_SIGNATURE = 'STATUS_WAITING_SIGNATURE',
  SIGNED = 'STATUS_SIGNED',
  NOT_SIGNED = 'STATUS_NOT_SIGNED',
  REFUSED = 'STATUS_REFUSED',
  ACCEPTED = 'STATUS_ACCEPTED',
  TRANSMITTED = 'STATUS_TRANSMITTED',
  SIGNATURE_IN_PROGRESS = 'STATUS_SIGNATURE_IN_PROGRESS',
}